requirements.txt
addict                 2.4.0
anyio                  3.5.0
argon2-cffi            21.3.0
argon2-cffi-bindings   21.2.0
attrs                  21.4.0
Babel                  2.9.1
backcall               0.2.0
beautifulsoup4         4.10.0
bleach                 4.1.0
brotlipy               0.7.0
certifi                2021.10.8
cffi                   1.14.6
chardet                4.0.0
conda                  4.10.3
conda-build            3.21.5
conda-package-handling 1.7.3
cryptography           3.4.8
cupy-cuda113           10.2.0
cycler                 0.11.0
debugpy                1.5.1
decorator              5.1.0
defusedxml             0.7.1
deprecation            2.1.0
dnspython              2.1.0
entrypoints            0.4
fastrlock              0.8
filelock               3.0.12
fonttools              4.29.1
glob2                  0.7
idna                   2.10
importlib-metadata     4.11.2
importlib-resources    5.4.0
ipykernel              6.9.1
ipython                7.27.0
ipython-genutils       0.2.0
ipywidgets             7.6.5
jedi                   0.18.0
Jinja2                 2.11.3
joblib                 1.1.0
json5                  0.9.6
jsonschema             4.4.0
jupyter-client         7.1.2
jupyter-core           4.9.2
jupyter-packaging      0.11.1
jupyter-server         1.13.5
jupyterlab             3.2.9
jupyterlab-pygments    0.1.2
jupyterlab-server      2.10.3
jupyterlab-widgets     1.0.2
kiwisolver             1.3.2
libarchive-c           2.9
MarkupSafe             2.0.1
matplotlib             3.5.1
matplotlib-inline      0.1.2
mistune                0.8.4
mkl-fft                1.3.1
mkl-random             1.2.2
mkl-service            2.4.0
nbclassic              0.3.5
nbclient               0.5.11
nbconvert              6.4.2
nbformat               5.1.3
nest-asyncio           1.5.4
notebook               6.4.8
numpy                  1.21.2
olefile                0.46
open3d                 0.14.1
opencv-contrib-python  4.5.5.62
opencv-python          4.5.5.62
opt-einsum             3.3.0
packaging              21.3
pandas                 1.3.5
pandocfilters          1.5.0
parso                  0.8.2
pexpect                4.8.0
pickleshare            0.7.5
Pillow                 8.4.0
pip                    21.0.1
pkginfo                1.7.1
plyfile                0.7.4
prometheus-client      0.13.1
prompt-toolkit         3.0.20
protobuf               3.19.4
psutil                 5.8.0
ptyprocess             0.7.0
pycosat                0.6.3
pycparser              2.20
Pygments               2.10.0
pynvrtc                9.2
pyOpenSSL              20.0.1
pyparsing              3.0.7
pyrsistent             0.18.1
PySocks                1.7.1
python-dateutil        2.8.2
python-etcd            0.4.5
pytz                   2021.3
PyYAML                 5.4.1
pyzmq                  22.3.0
requests               2.25.1
ruamel-yaml-conda      0.15.100
scikit-learn           1.0.2
scipy                  1.7.3
Send2Trash             1.8.0
setuptools             58.0.4
six                    1.16.0
sniffio                1.2.0
soupsieve              2.2.1
tensorboardX           2.5
terminado              0.13.1
testpath               0.6.0
threadpoolctl          3.1.0
tomlkit                0.10.0
torch                  1.10.0
torchelastic           0.2.0
torchtext              0.11.0
torchvision            0.11.0
tornado                6.1
tqdm                   4.61.2
traitlets              5.1.0
typing-extensions      3.10.0.2
urllib3                1.26.6
wcwidth                0.2.5
webencodings           0.5.1
websocket-client       1.3.1
wheel                  0.36.2
widgetsnbextension     3.5.2
yacs                   0.1.8
zipp                   3.7.0